moved all common components in separate folder and fixed all path issues
