+++ console/console-init/ui/src/components/Error/Error.tsx
