+++ console/console-init/ui/src/components/ErrorBoundary/ErrorBoundary.jsx
@@ -4,7 +4,7 @@
  */
 
 import React from "react";
-import { NotFound } from "./NotFound";
+import { NotFound } from "components";
 
 export class ErrorBoundary extends React.Component {
   constructor(props) {
