+++ console/console-init/ui/src/components/ErrorBoundary/index.ts
@@ -3,4 +3,4 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export {};
+export * from "./ErrorBoundary";
