+++ console/console-init/ui/src/components/NetworkStatusAlert/NetworkStatusAlert.tsx
