+++ console/console-init/ui/src/components/NoDataFound/NoDataFound.tsx
