+++ console/console-init/ui/src/components/WindowDimension/WindowDimension.tsx
@@ -13,7 +13,7 @@ function getWindowDimensions() {
   };
 }
 
-export default function useWindowDimensions() {
+export function useWindowDimensions() {
   const [windowDimensions, setWindowDimensions] = useState(
     getWindowDimensions()
   );
