--- console/console-init/ui/src/components/common/index.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright 2020, EnMasse authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-
-export * from "./About";
-export * from "./AddressSpaceListFormatter";
-export * from "./ConnectionListFormatter";
-export * from "./DialogPrompt";
-export * from "./Error";
-export * from "./ErrorBoundary";
-export * from "./NotFound";
-export * from "./ServerMessageAlert";
-export * from "./SortForMobileView";
-export * from "./WindowDimension";
-export * from "./NetworkStatusAlert";
