+++ console/console-init/ui/src/components/index.ts
@@ -7,3 +7,15 @@ export * from "./TablePagination";
 export * from "./TypeAheadSelect";
 export * from "./DropdownWithToggle";
 export * from "./DropdownWithKababToggle";
+export * from "./RootModal";
+export * from "./About";
+export * from "./DialogPrompt";
+export * from "./Error";
+export * from "./ErrorBoundary";
+export * from "./NotFound";
+export * from "./ServerMessageAlert";
+export * from "./SortForMobileView";
+export * from "./WindowDimension";
+export * from "./NetworkStatusAlert";
+export * from "./Toolbar";
+export * from "./NoDataFound";
