+++ console/console-init/ui/src/modules/address-detail/index.ts
@@ -1,1 +1,8 @@
-export {};
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+export * from "./components";
+export * from "./containers";
+export * from "./utils";
