+++ console/console-init/ui/src/modules/address-space/components/index.ts
@@ -10,3 +10,5 @@ export * from "./AddressSpaceConfiguration";
 export * from "./EditAddressSpace";
 export * from "./AddressSpaceReview";
 export * from "./AddressSpaceListKebab";
+export * from "./AddressSpaceHeader";
+export * from "./AddressSpaceNavigation";
