+++ console/console-init/ui/src/modules/address/index.ts
@@ -3,4 +3,8 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export {};
+export * from "./components";
+export * from "./containers";
+export * from "./dialogs";
+export * from "./utils";
+export * from "./AddressPage";
