+++ console/console-init/ui/src/modules/connection-detail/index.ts
@@ -1,1 +1,4 @@
 export * from "./ConnectionDetailPage";
+export * from "./components";
+export * from "./containers";
+export * from "./utils";
