+++ console/console-init/ui/src/modules/connection/containers/ConnectionContainer.tsx
@@ -5,14 +5,14 @@
 
 import React from "react";
 import { useQuery } from "@apollo/react-hooks";
-import { RETURN_ALL_CONECTION_LIST } from "graphql-module/queries";
 import { ISortBy } from "@patternfly/react-table";
+import { RETURN_ALL_CONECTION_LIST } from "graphql-module/queries";
 import {
   IConnection,
-  ConnectionList
-} from "modules/connection/components/ConnectionList/ConnectionList";
-import { EmptyConnection } from "modules/connection/components/EmptyConnection/EmptyConnection";
-import { getFilteredValue } from "components/common/ConnectionListFormatter";
+  ConnectionList,
+  EmptyConnection
+} from "modules/connection/components";
+import { getFilteredValue } from "utils";
 import { IConnectionListResponse } from "types/ResponseTypes";
 import { POLL_INTERVAL, FetchPolicy } from "constants/constants";
