+++ console/console-init/ui/src/utils/index.ts
@@ -1,1 +1,7 @@
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
 export * from "./utils";
+export * from "./Formatter";
